Module 3/project-smart-librarian/requirements.txt
openai
chromadb
python-dotenv
pydantic
rich
streamlit
pytest
requests>=2.32.0
langdetect
